Support to fetch pmon_daemon_control.json file from both platform and HWSKU folders (#18441)
- Why I did it
The pmon_daemon_control.json file currently has a platform-specific scope, resulting in all SKUs of the same platform sharing the same pmon daemon configuration. Consequently, this restricts the ability to have distinct pmon daemon configurations for different SKUs. The proposed change aims to enhance flexibility by allowing PMON to load this file from both the platform folder and the HWSKU folder. In case a pmon_daemon_control.json file exists in the HWSKU folder, it will take precedence over the one in the platform folder.

- How I did it
The pmon docker init script will be updated to prioritize the HWSKU folder when searching for the pmon_daemon_control.json file. If the file is present in the HWSKU folder, it will be utilized. However, if the file does not exist in the HWSKU folder, the script will then fallback to utilizing the one located in the platform folder.

- How to verify it
Build image with this change, put the file in
1. platform folder,
2. in HWSKU folder
3. in both folders
make sure the pmon daemons started according to the configuration.

Signed-off-by: Kebo Liu <kebol@nvidia.com>