Check and install sonic-platform package when pmon container is run every time
Originally, the sonic-platform package will only be installed during the creation of pmon.
If the switch is powered off during the creation and initialization of pmon,
the process to install sonic-platform package will be terminated and
the sonic-platform package will not be reinstalled when pmon container is restarted.

The sonic-platform package is the base module in pmon container.
The daemon can not be run correctlly if the sonic-platform package is not installed.

Modification:
- Remove the python version v2 and v3 check when generate supervisord.conf in pmon
- Installed the sonic-platform package if it is not existed when pmon container is run.
- Check the METADATA if the sonic_platform package has been installed.
  According to the description in the following link:
    https://packaging.python.org/en/latest/specifications/recording-installed-packages/#the-metadata-file
    The METADATA should be created after a package is installed and its size should not be zero.
- Let the related daemons be started after the sonic-platform package installation check is done.